requirements.txt
google-generativeai
sounddevice
numpy
customtkinter
python-dotenv
PyAudio
Pillow
keyboard
